+++ benihora/src/resample.rs
@@ -1,17 +1,17 @@
 #[derive(Debug, Clone)]
 pub enum Algo {
     UpSample {
-        in_per_out: f64,
-        prev_sample: f64,
-        next_sample: f64,
-        next_sample_time: f64,
+        in_per_out: f32,
+        prev_sample: f32,
+        next_sample: f32,
+        next_sample_time: f32,
     },
     DownSample {
-        in_per_out: f64,
-        out_per_in: f64,
-        left_value: f64,
-        right_value: f64,
-        time: f64,
+        in_per_out: f32,
+        out_per_in: f32,
+        left_value: f32,
+        right_value: f32,
+        time: f32,
     },
     Identity,
 }
@@ -21,7 +21,7 @@ pub struct Resample {
 }
 
 impl Resample {
-    pub fn new(input_sample_rate: f64, output_sample_rate: f64) -> Self {
+    pub fn new(input_sample_rate: f32, output_sample_rate: f32) -> Self {
         Self {
             algo: if input_sample_rate < output_sample_rate {
                 Algo::UpSample {
@@ -36,7 +36,7 @@ impl Resample {
                     out_per_in: output_sample_rate / input_sample_rate,
                     left_value: 0.0,
                     right_value: 0.0,
-                    time:0.0,
+                    time: 0.0,
                 }
             } else {
                 Algo::Identity
@@ -44,7 +44,7 @@ impl Resample {
         }
     }
 
-    pub fn process(&mut self, mut x: impl FnMut() -> f64) -> f64 {
+    pub fn process(&mut self, mut x: impl FnMut() -> f32) -> f32 {
         match self.algo {
             Algo::UpSample {
                 in_per_out,
